Fixed warnings for IPP9+ build;

+++ modules/imgproc/src/filter.cpp
@@ -4624,7 +4624,7 @@ static bool ipp_filter2D( InputArray _src, OutputArray _dst, int ddepth,
                     IppAutoBuffer<Ipp32f> kerTmp;
                     int kerStep = sizeof(Ipp32f)*kernelSize.width;
 #if IPP_VERSION_X100 >= 900
-                    if(kernel.step != kerStep)
+                    if((int)kernel.step != kerStep)
                     {
                         kerTmp.Alloc(kerStep*kernelSize.height);
                         if(ippiCopy_32f_C1R((Ipp32f*)kernel.data, (int)kernel.step, kerTmp, kerStep, kernelSize) < 0)
@@ -4651,7 +4651,7 @@ static bool ipp_filter2D( InputArray _src, OutputArray _dst, int ddepth,
                     IppAutoBuffer<Ipp16s> kerTmp;
                     int kerStep = sizeof(Ipp16s)*kernelSize.width;
 #if IPP_VERSION_X100 >= 900
-                    if(kernel.step != kerStep)
+                    if((int)kernel.step != kerStep)
                     {
                         kerTmp.Alloc(kerStep*kernelSize.height);
                         if(ippiCopy_16s_C1R((Ipp16s*)kernel.data, (int)kernel.step, kerTmp, kerStep, kernelSize) < 0)
